+++ 정훈/Chapter1_String/Chapter1_6.java
@@ -1,23 +1,20 @@
-package CodingTest_Study.정훈.Chapter1_String.Chapter1_String;
-
+package CodingTest_Study.정훈.Chapter1_String;
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
 
-public class Chapter1_1 {
+public class Chapter1_6 {
     public static void main(String[] args) throws Exception{
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         String str = br.readLine();
-        String s = br.readLine();
-        int answer = 0;
-        str = str.toLowerCase();
-        s = s.toLowerCase();
-        char temp = s.charAt(0);
+        String answer = "";
         for (int i = 0; i < str.length(); i++) {
-            if(temp == str.charAt(i)) {
-                answer++;
+            String temp = String.valueOf(str.charAt(i));
+            if(!answer.contains(temp)){
+                answer += temp;
             }
         }
+
         System.out.println(answer);
     }
 }
